+++ ros-20190910/alibabacloud_ros20190910/client.py
@@ -1500,6 +1500,8 @@ class Client(OpenApiClient):
     ) -> ros20190910_models.DeleteStackResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.delete_options):
+            query['DeleteOptions'] = request.delete_options
         if not UtilClient.is_unset(request.ram_role_name):
             query['RamRoleName'] = request.ram_role_name
         if not UtilClient.is_unset(request.region_id):
@@ -1536,6 +1538,8 @@ class Client(OpenApiClient):
     ) -> ros20190910_models.DeleteStackResponse:
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.delete_options):
+            query['DeleteOptions'] = request.delete_options
         if not UtilClient.is_unset(request.ram_role_name):
             query['RamRoleName'] = request.ram_role_name
         if not UtilClient.is_unset(request.region_id):